refactor: colocate invoice harvester skill runtime
