+++ skills/download-platform-invoices/adapters/anthropic.cjs
@@ -1,0 +1,4 @@
+module.exports = {
+  createAnthropicInvoiceAdapter:
+    require('./scrape.cjs').createAnthropicInvoiceAdapter,
+};
