+++ skills/download-platform-invoices/adapters/atlassian.cjs
@@ -1,0 +1,4 @@
+module.exports = {
+  createAtlassianInvoiceAdapter:
+    require('./scrape.cjs').createAtlassianInvoiceAdapter,
+};
