+++ skills/download-platform-invoices/adapters/gcp.cjs
@@ -1,0 +1,3 @@
+module.exports = {
+  createGcpInvoiceAdapter: require('./scrape.cjs').createGcpInvoiceAdapter,
+};
