+++ skills/download-platform-invoices/adapters/github.cjs
@@ -1,0 +1,4 @@
+module.exports = {
+  createGitHubInvoiceAdapter:
+    require('./scrape.cjs').createGitHubInvoiceAdapter,
+};
