+++ skills/download-platform-invoices/adapters/linkedin.cjs
@@ -1,0 +1,4 @@
+module.exports = {
+  createLinkedInInvoiceAdapter:
+    require('./scrape.cjs').createLinkedInInvoiceAdapter,
+};
