+++ skills/download-platform-invoices/adapters/openai.cjs
@@ -1,0 +1,4 @@
+module.exports = {
+  createOpenAIInvoiceAdapter:
+    require('./scrape.cjs').createOpenAIInvoiceAdapter,
+};
