+++ skills/download-platform-invoices/fixtures/recorded-anthropic.json
