+++ skills/download-platform-invoices/fixtures/recorded-atlassian.json
