+++ skills/download-platform-invoices/fixtures/recorded-azure.json
