+++ skills/download-platform-invoices/fixtures/recorded-gcp.json
