+++ skills/download-platform-invoices/fixtures/recorded-google-ads.json
