+++ skills/download-platform-invoices/fixtures/recorded-openai.json
