+++ skills/download-platform-invoices/fixtures/recorded-stripe.json
