+++ skills/download-platform-invoices/fixtures/stripe-invoices.json
