+++ skills/download-platform-invoices/schema.json
@@ -1,0 +1,37 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "urn:hybridclaw:schema:invoice-record",
+  "title": "HybridClaw normalized invoice record",
+  "type": "object",
+  "additionalProperties": false,
+  "required": [
+    "vendor",
+    "invoice_no",
+    "period",
+    "issue_date",
+    "due_date",
+    "net",
+    "vat_rate",
+    "vat",
+    "gross",
+    "currency",
+    "pdf_path",
+    "source_url",
+    "checksum_sha256"
+  ],
+  "properties": {
+    "vendor": { "type": "string", "minLength": 1 },
+    "invoice_no": { "type": "string", "minLength": 1 },
+    "period": { "type": "string", "pattern": "^\\d{4}-\\d{2}$" },
+    "issue_date": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
+    "due_date": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
+    "net": { "type": "number" },
+    "vat_rate": { "type": "number", "minimum": 0, "maximum": 1 },
+    "vat": { "type": "number" },
+    "gross": { "type": "number" },
+    "currency": { "type": "string", "pattern": "^[A-Z]{3}$" },
+    "pdf_path": { "type": "string", "minLength": 1 },
+    "source_url": { "type": "string", "minLength": 1 },
+    "checksum_sha256": { "type": "string", "pattern": "^[a-f0-9]{64}$" }
+  }
+}
