--- src/invoices/index.ts
@@ -1,18 +1,0 @@
-export * from './adapters/api-client.js';
-export * from './adapters/dashboard-scrape.js';
-export * from './adapters/playwright-scrape.js';
-export * from './adapters/recorded-fixture.js';
-export * from './adapters/stripe.js';
-export * from './audit.js';
-export * from './config.js';
-export * from './credentials.js';
-export * from './date-utils.js';
-export * from './datev-upload.js';
-export * from './harvester.js';
-export * from './idempotency.js';
-export * from './monthly-run.js';
-export * from './provider-registry.js';
-export * from './schema.js';
-export * from './schema-error.js';
-export * from './scrape-plans.js';
-export * from './types.js';
